Rename "quads" to "placeShapedText" / "placeShapedIcons"
