+++ js/symbol/symbol_quads_array.js
@@ -3,10 +3,6 @@
 var StructArrayType = require('../util/struct_array');
 var util = require('../util/util');
 var Point = require('point-geometry');
-var SymbolQuad = require('./quads').SymbolQuad;
-
-// notes from ansis on slack:
-// it would be best if they are added to a buffer in advance so that they are only created once. There would be a separate buffer with all the individual collision boxes and then SymbolInstance would store the beginning and end indexes of a feature's collisionboxes. CollisionFeature wouldn't really exist as a standalone thing, it would just be a range of boxes in the big collision box buffer
 
 /*
  *
@@ -57,16 +53,17 @@ util.extendAll(SymbolQuadsArray.prototype.StructType.prototype, {
         return new Point(this.anchorPointX, this.anchorPointY);
     },
     get SymbolQuad() {
-        return new SymbolQuad(this.anchorPoint,
-            new Point(this.tlX, this.tlY),
-            new Point(this.trX, this.trY),
-            new Point(this.blX, this.blY),
-            new Point(this.brX, this.brY),
-            { x: this.texX, y: this.texY, h: this.texH, w: this.texW, height: this.texH, width: this.texW },
-            this.anchorAngle,
-            this.glyphAngle,
-            this.minScale,
-            this.maxScale);
+        return {
+            anchorPoint: this.anchorPoint,
+            tl: new Point(this.tlX, this.tlY),
+            tr: new Point(this.trX, this.trY),
+            bl: new Point(this.blX, this.blY),
+            br: new Point(this.brX, this.brY),
+            tex: { x: this.texX, y: this.texY, h: this.texH, w: this.texW, height: this.texH, width: this.texW },
+            anchorAngle: this.anchorAngle,
+            glyphAngle: this.glyphAngle,
+            minScale: this.minScale,
+            maxScale: this.maxScale
+        };
     }
 });
-
