+++ test/js/data/symbol_bucket.test.js
@@ -9,7 +9,7 @@ var SymbolBucket = require('../../../js/data/bucket/symbol_bucket');
 var Collision = require('../../../js/symbol/collision_tile');
 var CollisionBoxArray = require('../../../js/symbol/collision_box');
 var SymbolInstancesArray = require('../../../js/symbol/symbol_instances');
-var SymbolQuadsArray = require('../../../js/symbol/symbol_quads');
+var SymbolQuadsArray = require('../../../js/symbol/symbol_quads_array');
 var GlyphAtlas = require('../../../js/symbol/glyph_atlas');
 var StyleLayer = require('../../../js/style/style_layer');
 var util = require('../../../js/util/util');
