+++ test/js/symbol/place_shaped_icon.test.js
@@ -2,7 +2,7 @@
 
 var test = require('mapbox-gl-js-test').test;
 
-var getIconQuads = require('../../../js/symbol/quads').getIconQuads;
+var placeShapedIcons = require('../../../js/symbol/place_shaped_icon');
 var Anchor = require('../../../js/symbol/anchor');
 var Point = require('point-geometry');
 
@@ -28,14 +28,14 @@ function createShapedIcon() {
     };
 }
 
-test('getIconQuads', function(t) {
+test('placeShapedIcons', function(t) {
 
     t.test('point', function(t) {
         var anchor = new Anchor(2, 3, 0, undefined);
         var layer = createLayer({
             layout: {'icon-rotate': 0}
         });
-        t.deepEqual(getIconQuads(anchor, createShapedIcon(), 2, [], layer, false), [
+        t.deepEqual(placeShapedIcons(anchor, createShapedIcon(), 2, [], layer, false), [
             {
                 anchorPoint: { x: 2, y: 3 },
                 tl: { x: -8, y: -6 },
@@ -55,7 +55,7 @@ test('getIconQuads', function(t) {
         var layer = createLayer({
             layout: {'icon-rotate': 0}
         });
-        t.deepEqual(getIconQuads(anchor, createShapedIcon(), 2, [new Point(0, 0), new Point(8, 9)], layer, false), [
+        t.deepEqual(placeShapedIcons(anchor, createShapedIcon(), 2, [new Point(0, 0), new Point(8, 9)], layer, false), [
             {
                 anchorPoint: { x: 2, y: 3},
                 tl: { x: -8, y: -6 },
@@ -72,7 +72,7 @@ test('getIconQuads', function(t) {
     t.end();
 });
 
-test('getIconQuads text-fit', function(t) {
+test('placeShapedIcons text-fit', function(t) {
     var anchor = new Anchor(0, 0, 0, undefined);
     function createShapedIcon() {
         return {
@@ -87,7 +87,7 @@ test('getIconQuads text-fit', function(t) {
         };
     }
 
-    function createshapedText() {
+    function createShapedText() {
         return {
             top: -10,
             bottom: 30,
@@ -97,22 +97,22 @@ test('getIconQuads text-fit', function(t) {
     }
 
     t.test('icon-text-fit: none', function(t) {
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'icon-text-fit': 'none'
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -11, y: -11 });
         t.deepEqual(quads[0].tr, { x: 9, y: -11 });
         t.deepEqual(quads[0].bl, { x: -11, y: 9 });
         t.deepEqual(quads[0].br, { x: 9, y: 9 });
 
-        t.deepEqual(quads, getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        t.deepEqual(quads, placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'icon-text-fit': 'none',
                 'icon-text-fit-padding': [10, 10]
             }
-        }), false, createshapedText()), 'ignores padding');
+        }), false, createShapedText()), 'ignores padding');
 
         t.end();
     });
@@ -120,13 +120,13 @@ test('getIconQuads text-fit', function(t) {
     t.test('icon-text-fit: width', function(t) {
         // - Uses text width
         // - Preserves icon height, centers vertically
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 24,
                 'icon-text-fit': 'width',
                 'icon-text-fit-padding': [ 0, 0, 0, 0 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -60, y: 0 });
         t.deepEqual(quads[0].tr, { x: 20, y: 0 });
         t.deepEqual(quads[0].bl, { x: -60, y: 20 });
@@ -137,13 +137,13 @@ test('getIconQuads text-fit', function(t) {
     t.test('icon-text-fit: width, x textSize', function(t) {
         // - Uses text width (adjusted for textSize)
         // - Preserves icon height, centers vertically
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 12,
                 'icon-text-fit': 'width',
                 'icon-text-fit-padding': [ 0, 0, 0, 0 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -30, y: -5 });
         t.deepEqual(quads[0].tr, { x: 10, y: -5 });
         t.deepEqual(quads[0].bl, { x: -30, y: 15 });
@@ -155,13 +155,13 @@ test('getIconQuads text-fit', function(t) {
         // - Uses text width (adjusted for textSize)
         // - Preserves icon height, centers vertically
         // - Applies padding x, padding y
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 12,
                 'icon-text-fit': 'width',
                 'icon-text-fit-padding': [ 5, 10, 5, 10 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -40, y: -10 });
         t.deepEqual(quads[0].tr, { x: 20, y: -10 });
         t.deepEqual(quads[0].bl, { x: -40, y: 20 });
@@ -172,13 +172,13 @@ test('getIconQuads text-fit', function(t) {
     t.test('icon-text-fit: height', function(t) {
         // - Uses text height
         // - Preserves icon width, centers horizontally
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 24,
                 'icon-text-fit': 'height',
                 'icon-text-fit-padding': [ 0, 0, 0, 0 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -30, y: -10 });
         t.deepEqual(quads[0].tr, { x: -10, y: -10 });
         t.deepEqual(quads[0].bl, { x: -30, y: 30 });
@@ -189,13 +189,13 @@ test('getIconQuads text-fit', function(t) {
     t.test('icon-text-fit: height, x textSize', function(t) {
         // - Uses text height (adjusted for textSize)
         // - Preserves icon width, centers horizontally
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 12,
                 'icon-text-fit': 'height',
                 'icon-text-fit-padding': [ 0, 0, 0, 0 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -20, y: -5 });
         t.deepEqual(quads[0].tr, { x: 0, y: -5 });
         t.deepEqual(quads[0].bl, { x: -20, y: 15 });
@@ -207,13 +207,13 @@ test('getIconQuads text-fit', function(t) {
         // - Uses text height (adjusted for textSize)
         // - Preserves icon width, centers horizontally
         // - Applies padding x, padding y
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 12,
                 'icon-text-fit': 'height',
                 'icon-text-fit-padding': [ 5, 10, 5, 10 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -30, y: -10 });
         t.deepEqual(quads[0].tr, { x: 10, y: -10 });
         t.deepEqual(quads[0].bl, { x: -30, y: 20 });
@@ -223,13 +223,13 @@ test('getIconQuads text-fit', function(t) {
 
     t.test('icon-text-fit: both', function(t) {
         // - Uses text width + height
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 24,
                 'icon-text-fit': 'both',
                 'icon-text-fit-padding': [ 0, 0, 0, 0 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -60, y: -10 });
         t.deepEqual(quads[0].tr, { x: 20, y: -10 });
         t.deepEqual(quads[0].bl, { x: -60, y: 30 });
@@ -239,13 +239,13 @@ test('getIconQuads text-fit', function(t) {
 
     t.test('icon-text-fit: both, x textSize', function(t) {
         // - Uses text width + height (adjusted for textSize)
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 12,
                 'icon-text-fit': 'both',
                 'icon-text-fit-padding': [ 0, 0, 0, 0 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -30, y: -5 });
         t.deepEqual(quads[0].tr, { x: 10, y: -5 });
         t.deepEqual(quads[0].bl, { x: -30, y: 15 });
@@ -256,13 +256,13 @@ test('getIconQuads text-fit', function(t) {
     t.test('icon-text-fit: both, x textSize, + padding', function(t) {
         // - Uses text width + height (adjusted for textSize)
         // - Applies padding x, padding y
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 12,
                 'icon-text-fit': 'both',
                 'icon-text-fit-padding': [ 5, 10, 5, 10 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -40, y: -10 });
         t.deepEqual(quads[0].tr, { x: 20, y: -10 });
         t.deepEqual(quads[0].bl, { x: -40, y: 20 });
@@ -273,13 +273,13 @@ test('getIconQuads text-fit', function(t) {
     t.test('icon-text-fit: both, padding t/r/b/l', function(t) {
         // - Uses text width + height (adjusted for textSize)
         // - Applies padding t/r/b/l
-        var quads = getIconQuads(anchor, createShapedIcon(), 2, [], createLayer({
+        var quads = placeShapedIcons(anchor, createShapedIcon(), 2, [], createLayer({
             layout: {
                 'text-size': 12,
                 'icon-text-fit': 'both',
                 'icon-text-fit-padding': [ 0, 5, 10, 15 ]
             }
-        }), false, createshapedText());
+        }), false, createShapedText());
         t.deepEqual(quads[0].tl, { x: -45, y: -5 });
         t.deepEqual(quads[0].tr, { x: 15, y: -5 });
         t.deepEqual(quads[0].bl, { x: -45, y: 25 });
